@@ -24,8 +24,8 @@
     "node-uuid": "1.4.8",
     "prop-types": "15.6.2",
     "pug": "2.0.0-rc.1",
-    "react": "15.0.1",
-    "react-dom": "15.0.1",
+    "react": "15.3.0",
+    "react-dom": "15.3.0",
     "react-fa": "4.0.1",
     "uuid": "2.0.2"
   },
