@@ -1,4 +1,5 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import GnomeView from '../gnome-view/gnome-view';
 import LoadingSpinner from '../loading-spinner/loading-spinner';
 import './gnome-list-view.css';
@@ -20,6 +21,6 @@ export default class GnomeListView extends React.Component {
 };
 
 GnomeListView.propTypes = {
-    createExpandItemAction: React.PropTypes.func,
-    items: React.PropTypes.arrayOf(React.PropTypes.object).isRequired
+    createExpandItemAction: PropTypes.func,
+    items: PropTypes.arrayOf(PropTypes.object).isRequired
 }
