@@ -1,5 +1,6 @@
 import React from 'react';
 import Icon from 'react-fa';
+import PropTypes from 'prop-types';
 import { createAddToCartAction } from '../../actions/cart-actions';
 
 import './gnome-view.css';
@@ -9,11 +10,11 @@ export default class GnomeView extends React.Component {
         super(props, context);
     }
 
-    async onAddToCartClicked() {
+    onAddToCartClicked = async () => {
         await createAddToCartAction(this.props.item);
     }
 
-    onCoverartClicked() {
+    onCoverartClicked = () => {
         this.props.createExpandItemAction(this.props.item.id);
     }
 
@@ -48,6 +49,6 @@ export default class GnomeView extends React.Component {
 };
 
 GnomeView.propTypes = {
-    createExpandItemAction: React.PropTypes.func.isRequired,
-    item: React.PropTypes.object.isRequired
+    createExpandItemAction: PropTypes.func.isRequired,
+    item: PropTypes.object.isRequired
 }
